fix: moved validateInput method to separate file and added some additional validation
